requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# \canvas-todo-linker\to_do_connector.py: 6,9
msal == 1.5.0

# \canvas-todo-linker\canvas_connector.py: 1,2
# \canvas-todo-linker\canvas_todo_linker.py: 1
# \canvas-todo-linker\to_do_connector.py: 8,10
requests == 2.23.0

python-dateutil
